Migrate FigmaDocInfo and EmbeddedImageMap
Towards: Complete migration of DesignComposeDefinition to dc_bundle. #1457
