+++ .gitignore
@@ -132,3 +132,6 @@ fabric.properties
 # Android studio 3.1+ serialized cache file
 .idea/caches
 .idea/runConfigurations.xml
+
+# Other
+.idea/kotlinc.xml
