+++ crates/dc_bundle/Cargo.toml
@@ -11,10 +11,10 @@ rust-version.workspace = true
 prost.workspace = true
 thiserror.workspace = true
 serde.workspace = true
+serde_bytes.workspace = true
 
 # Temporary dependencies during the transition to protobuf structures
 euclid.workspace = true
-
 
 [build-dependencies]
 prost-build.workspace = true
