+++ crates/dc_bundle/src/legacy_definition.rs
@@ -16,13 +16,30 @@
 
 // This module holds the original structures that made up the serialized design doc, and which will be replaced with the protobuf implementations of the Design Definition
 
+use serde::{Deserialize, Serialize};
+use std::collections::HashMap;
+use std::sync::Arc;
 // To help keep the legacy definition files clear we alias `crate::definition`, which is the base
 // module for the generated protobuf files to `proto`, so that all of the protobuf-generated types
 // inside `legacy_definition` must be prepended with `proto::`
 pub(crate) use crate::definition as proto;
+use crate::legacy_definition::element::background::ImageKey;
+
 pub mod element;
 pub mod interaction;
 pub mod layout;
 pub mod modifier;
 pub mod plugin;
 pub mod view;
+
+/// EncodedImageMap contains a mapping from ImageKey to network bytes. It can create an
+/// ImageMap and is intended to be used when we want to use Figma-defined components but do
+/// not want to communicate with the Figma service.
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct EncodedImageMap(pub HashMap<ImageKey, Arc<serde_bytes::ByteBuf>>);
+
+impl EncodedImageMap {
+    pub fn map(&self) -> HashMap<ImageKey, Arc<serde_bytes::ByteBuf>> {
+        self.0.clone()
+    }
+}
